Use TalonSRX Follower mode instead of WPILib SpeedControllerGroups
This uses the built in Talon SRX follower mode to copy the other Talons' output. This prepares for more advanced features later, like controlling the left and right sides with PID (i.e. PID to a set velocity or position), and probably also lightens the load just a tiny bit on the RIO (O(N) to O(1) where N is the number of controllers in the group?), as following is handled by the Talons themselves on their hardware instead of the for loop running on the RIO from SpeedControllerGroup.