Do some work
Okay... this is kind of confusing. The robot now can move forwards/back
to a specified distance (tested w/ 1m)... without velocity PID.

So... we can a) change it to use velocity PID (not recommended... it
actually works now), b) work on AutoTurnPIDCommand, as it doesn't work
yet, or c) try to implement motion paths. We also need to start working
on code for game mechs.